Add renamed files
